# backend/docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ML Evaluation Platform API",
    "description": "API for object detection and segmentation model evaluation platform",
    "version": "1.0.0",
    "contact": {
      "name": "ML Evaluation Platform"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.ml-eval.cloud",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/training/jobs": {
      "post": {
        "summary": "Start model training",
        "description": "Initiate fine-tuning of a model with user annotations",
        "operationId": "startTraining",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Training job created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJob"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/training/jobs/{job_id}": {
      "get": {
        "summary": "Get training job status",
        "description": "Monitor training job progress and logs",
        "operationId": "getTrainingJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Training job status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJob"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TrainingRequest": {
        "type": "object",
        "properties": {
          "base_model_id": {
            "type": "string",
            "format": "uuid"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid"
          },
          "hyperparameters": {
            "type": "object",
            "properties": {
              "epochs": {
                "type": "integer",
                "default": 100
              },
              "batch_size": {
                "type": "integer",
                "default": 16
              },
              "learning_rate": {
                "type": "number",
                "format": "float",
                "default": 0.001
              },
              "patience": {
                "type": "integer",
                "default": 10
              }
            },
            "required": [
              "epochs",
              "batch_size",
              "learning_rate"
            ]
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "base_model_id",
          "dataset_id",
          "hyperparameters"
        ]
      },
      "TrainingJob": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "base_model_id": {
            "type": "string",
            "format": "uuid"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ]
          },
          "progress_percentage": {
            "type": "number",
            "format": "float"
          },
          "hyperparameters": {
            "type": "object"
          },
          "execution_logs": {
            "type": "string"
          },
          "start_time": {
            "type": "string",
            "format": "date-time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time"
          },
          "result_model_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "base_model_id",
          "dataset_id",
          "status",
          "progress_percentage",
          "hyperparameters",
          "created_at"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "error",
          "message",
          "timestamp"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}